pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "utils_cmip7"
version = "0.4.0"
description = "Carbon cycle analysis toolkit for Unified Model outputs"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Yousheng Li", email = "ysli.13477426@gmail.com"}
]
keywords = ["climate", "carbon-cycle", "CMIP", "UM", "NetCDF"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Atmospheric Science",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20,<1.25",  # Pin to avoid binary incompatibility on Python 3.9
    "pandas>=1.3",
    "matplotlib>=3.4",
    "scitools-iris>=3.0,<3.9",  # Pin <3.9 to maintain Python 3.9 support
    "cartopy>=0.20",
    "xarray>=0.19",
    "cf-units>=2.1,<4.0",  # Pin major version for stability
    "cftime>=1.5,<1.7",  # Pin for Python 3.8/3.9 compatibility
    "netCDF4>=1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "flake8>=4.0",
    "black>=22.0",
    "isort>=5.10",
    "mypy>=0.950",
]
docs = [
    "sphinx>=4.5",
    "sphinx-rtd-theme>=1.0",
]
param-viz = [
    "scikit-learn>=1.0",
]

[project.urls]
Homepage = "https://github.com/Climateyousheng/utils_cmip7"
Repository = "https://github.com/Climateyousheng/utils_cmip7"
Issues = "https://github.com/Climateyousheng/utils_cmip7/issues"

[project.scripts]
utils-cmip7-extract-raw = "utils_cmip7.cli:extract_raw_cli"
utils-cmip7-extract-preprocessed = "utils_cmip7.cli:extract_preprocessed_cli"
utils-cmip7-validate-experiment = "utils_cmip7.cli:validate_experiment_cli"
utils-cmip7-validate-ppe = "utils_cmip7.cli:validate_ppe_cli"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
utils_cmip7 = ["data/obs/*.csv"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "-v",
    "--cov=src/utils_cmip7",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["src/utils_cmip7"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
